feat: 重构工具注册逻辑，保证架构脉络整洁清晰，添加
